Abstract away mapping objects to make form config easier
