+++ src/form/birth/required-fields.ts
@@ -22,6 +22,8 @@ import {
   placeOfBirthOptions
 } from '../common/select-options'
 import { certificateHandlebars } from './certificate-handlebars'
+import { getFieldMapping } from '@countryconfig/utils/mapping/field-mapping-utils'
+import { getEventLocationSelectionMapping } from '@countryconfig/utils/address-utils'
 
 export const informantType: SerializedFormField = {
   name: 'informantType',
@@ -32,20 +34,10 @@ export const informantType: SerializedFormField = {
   initialValue: '',
   validator: [],
   placeholder: formMessageDescriptors.formSelectPlaceholder,
-  mapping: {
-    mutation: {
-      operation: 'fieldValueSectionExchangeTransformer',
-      parameters: ['registration', 'informantType']
-    },
-    query: {
-      operation: 'fieldValueSectionExchangeTransformer',
-      parameters: ['registration', 'informantType']
-    },
-    template: {
-      fieldName: certificateHandlebars.informantType,
-      operation: 'selectTransformer'
-    }
-  },
+  mapping: getFieldMapping(
+    'informantType',
+    certificateHandlebars.informantType
+  ),
   options: birthInformantTypeOptions
 }
 
@@ -71,16 +63,7 @@ export const getPlaceOfBirthFields = (): SerializedFormField[] => [
     validator: [],
     placeholder: formMessageDescriptors.formSelectPlaceholder,
     options: placeOfBirthOptions,
-    mapping: {
-      mutation: {
-        operation: 'eventLocationMutationTransformer',
-        parameters: [{ useCase: 'placeOfBirth' }]
-      },
-      query: {
-        operation: 'eventLocationTypeQueryTransformer',
-        parameters: []
-      }
-    }
+    mapping: getEventLocationSelectionMapping('placeOfBirth')
   },
   {
     name: 'birthLocation',
@@ -105,21 +88,10 @@ export const getPlaceOfBirthFields = (): SerializedFormField[] => [
         expression: '(values.placeOfBirth!="HEALTH_FACILITY")'
       }
     ],
-    mapping: {
-      template: {
-        fieldName: certificateHandlebars.placeOfBirth,
-        operation: 'eventLocationNameQueryOfflineTransformer',
-        parameters: ['facilities', 'placeOfBirth']
-      },
-      mutation: {
-        operation: 'eventLocationMutationTransformer',
-        parameters: [{ useCase: 'placeOfBirth' }]
-      },
-      query: {
-        operation: 'eventLocationIDQueryTransformer',
-        parameters: []
-      }
-    }
+    mapping: getEventLocationSelectionMapping(
+      'birthLocation',
+      certificateHandlebars.placeOfBirth
+    )
   }
 ]
 
@@ -138,11 +110,7 @@ export const getDetailsExist = (
     initialValue: true,
     validator: [],
     conditionals,
-    mapping: {
-      query: {
-        operation: 'booleanTransformer'
-      }
-    }
+    mapping: getFieldMapping('detailsExist')
   } satisfies SerializedFormField)
 
 export const getReasonNotExisting = (certificateHandlebar: string) =>
@@ -159,10 +127,5 @@ export const getReasonNotExisting = (certificateHandlebar: string) =>
     validator: [],
     initialValue: '',
     required: true,
-    mapping: {
-      template: {
-        fieldName: certificateHandlebar,
-        operation: 'plainInputTransformer'
-      }
-    }
+    mapping: getFieldMapping('reasonNotApplying', certificateHandlebar)
   } satisfies SerializedFormField)
