+++ src/form/custom-fields.ts
@@ -10,8 +10,8 @@
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
 
-import { createCustomFieldHandlebarName } from '@countryconfig/utils'
 import { SerializedFormField } from './types/types'
+import { getCustomFieldMapping } from '@countryconfig/utils/mapping/field-mapping-utils'
 
 // ======================= CUSTOM FIELD CONFIGURATION =======================
 
@@ -19,44 +19,30 @@ import { SerializedFormField } from './types/types'
 // DUPLICATE AND RENAME FUNCTIONS LIKE THESE IN ORDER TO USE SIMILAR FIELDS
 
 export function createCustomFieldExample(): SerializedFormField {
-  // THE fieldId STRING IS A DOT SEPARATED STRING AND IS IMPORTANT TO SET CORRECTLY
+  // GIVE THE FIELD A UNIQUE NAME.  IF THE NAME IS ALREADY IN USE, YOU WILL NOTICE AN ERROR ON PAGE LOAD IN DEVELOPMENT
+  const fieldName: string = 'favoriteColor'
+  // THE fieldId STRING IS A DOT SEPARATED STRING AND IS IMPORTANT TO SET CORRECTLY DEPENDING ON WHERE THE CUSTOM FIELD IS LOCATED
   // THE FORMAT IS event.sectionId.groupId.uniqueFieldName
-  const fieldId: string = 'birth.child.child-view-group.favouriteColor'
-
-  // THE HANDLEBAR IS CREATED BY THIS FUNCTION IN ORDER TO USE THE VALUE ON A CERTIFICATE
-  const customFieldCertificateHandlebar =
-    createCustomFieldHandlebarName(fieldId)
-  console.log(
-    'Custom field addded with handlebar: ',
-    customFieldCertificateHandlebar
-  ) // WILL RESOLVE TO "{{birthChildFavouriteColor}}"
+  const fieldId: string = `birth.child.child-view-group.${fieldName}`
+  // IN ORDER TO USE THE VALUE ON A CERTIFICATE
+  // THE groupId IS IGNORED AND THE HANDLEBAR WILL LOG IN THE CONSOLE
+  // IN THIS EXAMPLE, IT WILL RESOLVE IN CAMELCASE TO "{{birthChildFavouriteColor}}"
 
   return {
-    name: 'favoriteColor',
+    name: fieldName,
     customQuesstionMappingId: fieldId,
     custom: true,
     required: true,
-    type: 'TEXT',
+    type: 'TEXT', // ANY FORM FIELD TYPE IS POSSIBLE. ADD ADDITIONAL PROPS AS REQUIRED.  REFER TO THE form/README.md FILE
     label: {
       id: 'form.customField.label.favoriteColor',
       description: 'A form field that asks for the persons favorite color.',
       defaultMessage: 'What is your favorite color?'
     },
     initialValue: '',
-    validator: [],
-    mapping: {
-      mutation: {
-        operation: 'customFieldToQuestionnaireTransformer'
-      },
-      query: {
-        operation: 'questionnaireToCustomFieldTransformer'
-      },
-      template: {
-        fieldName: customFieldCertificateHandlebar,
-        operation: 'questionnaireToTemplateFieldTransformer'
-      }
-    },
-    conditionals: [],
+    validator: [], // EDIT VALIDATORS AS YOU SEE FIT
+    mapping: getCustomFieldMapping(fieldId), // ALL CUSTOM FIELDS MUST USE THIS MAPPING FUNCTION
+    conditionals: [], // EDIT VALIDATORS AS YOU SEE FIT
     maxLength: 250
   }
 }
