+++ src/form/marriage/index.ts
@@ -58,12 +58,25 @@ import {
   witnessTwoNameInEnglish
 } from '../common/preview-groups'
 import { certificateHandlebars } from './certificate-handlebars'
+import { getCommonSectionMapping } from '@countryconfig/utils/mapping/field-mapping-utils'
+
+// import { createCustomFieldExample } from '../custom-fields'
 
 // ======================= FORM CONFIGURATION =======================
 
 // A REGISTRATION FORM IS MADE UP OF PAGES OR "SECTIONS"
 
-// A "SECTION" CAN BE SPLIT OVER MULTIPLE PAGES USING "GROUPS" ALTHOUGH THIS MAY BE DEPRECATED
+// A "SECTION" CAN BE SPLIT OVER MULTIPLE SUB-PAGES USING "GROUPS"
+
+// GROUPS CONTAIN A FIELDS ARRAY AND EACH FIELD IS RENDERED BY A FORM FIELD FUNCTION
+
+// MOVE FORM FIELD FUNCTIONS UP AND DOWN TO CHANGE THE VERTICAL ORDER OF FIELDS
+
+// IN EACH GROUP, REQUIRED FIELDS MUST BE INCLUDED AS-IS FOR OPENCRVS TO FUNCTION
+
+// OPTIONAL FIELDS CAN BE COMMENTED OUT OR REMOVED IF NOT REQUIRED
+
+// DUPLICATE & FOLLOW THE INSTRUCTIONS IN THE createCustomFieldExample FUNCTION WHEN REQUIRED FOR ADDING NEW CUSTOM FIELDS
 
 export const marriageForm: ISerializedForm = {
   sections: [
@@ -129,14 +142,7 @@ export const marriageForm: ISerializedForm = {
           previewGroups: [informantNameInEnglish]
         }
       ],
-      mapping: {
-        mutation: {
-          operation: 'setInformantSectionTransformer'
-        },
-        query: {
-          operation: 'getInformantSectionTransformer'
-        }
-      }
+      mapping: getCommonSectionMapping('informant')
     },
     {
       id: 'groom',
@@ -268,14 +274,7 @@ export const marriageForm: ISerializedForm = {
           previewGroups: [witnessOneNameInEnglish]
         }
       ],
-      mapping: {
-        mutation: {
-          operation: 'setInformantSectionTransformer'
-        },
-        query: {
-          operation: 'getInformantSectionTransformer'
-        }
-      }
+      mapping: getCommonSectionMapping('informant')
     },
     {
       id: 'witnessTwo',
@@ -302,14 +301,7 @@ export const marriageForm: ISerializedForm = {
           previewGroups: [witnessTwoNameInEnglish]
         }
       ],
-      mapping: {
-        mutation: {
-          operation: 'setInformantSectionTransformer'
-        },
-        query: {
-          operation: 'getInformantSectionTransformer'
-        }
-      }
+      mapping: getCommonSectionMapping('informant')
     },
     documentsSection
   ]
