+++ src/form/marriage/required-sections.ts
@@ -1,7 +1,7 @@
+import { getSectionMapping } from '@countryconfig/utils/mapping/section/marriage/mapping-utils'
 import { getInformantConditionalForMarriageDocUpload } from '../common/default-validation-conditionals'
 import { formMessageDescriptors } from '../common/messages'
 import { ISelectOption, ISerializedFormSection } from '../types/types'
-import { certificateHandlebars } from './certificate-handlebars'
 import { getDocUploaderForMarriage } from './required-fields'
 
 export const registrationSection = {
@@ -13,76 +13,7 @@ export const registrationSection = {
     id: 'form.section.declaration.name'
   },
   groups: [],
-  mapping: {
-    template: [
-      {
-        fieldName: certificateHandlebars.registrationNumber,
-        operation: 'registrationNumberTransformer'
-      },
-      {
-        fieldName: certificateHandlebars.qrCode,
-        operation: 'QRCodeTransformer'
-      },
-      {
-        fieldName: certificateHandlebars.certificateDate,
-        operation: 'certificateDateTransformer',
-        parameters: ['en', 'dd MMMM yyyy']
-      },
-      {
-        fieldName: certificateHandlebars.registrar,
-        operation: 'userTransformer',
-        parameters: ['REGISTERED']
-      },
-      {
-        fieldName: certificateHandlebars.registrationAgent,
-        operation: 'userTransformer',
-        parameters: ['VALIDATED']
-      },
-      {
-        fieldName: certificateHandlebars.registrarName,
-        operation: 'registrarNameUserTransformer'
-      },
-      {
-        fieldName: certificateHandlebars.role,
-        operation: 'roleUserTransformer'
-      },
-      {
-        fieldName: certificateHandlebars.registrarSignature,
-        operation: 'registrarSignatureUserTransformer'
-      },
-      {
-        fieldName: certificateHandlebars.registrationDate,
-        operation: 'registrationDateTransformer',
-        parameters: ['en', 'dd MMMM yyyy']
-      },
-      {
-        fieldName: certificateHandlebars.registrationLocation,
-        operation: 'registrationLocationUserTransformer'
-      },
-      {
-        fieldName: certificateHandlebars.groomSignature,
-        operation: 'groomSignatureTransformer'
-      },
-      {
-        fieldName: certificateHandlebars.brideSignature,
-        operation: 'brideSignatureTransformer'
-      },
-      {
-        fieldName: certificateHandlebars.witnessOneSignature,
-        operation: 'witnessOneSignatureTransformer'
-      },
-      {
-        fieldName: certificateHandlebars.witnessTwoSignature,
-        operation: 'witnessTwoSignatureTransformer'
-      }
-    ],
-    mutation: {
-      operation: 'setMarriageRegistrationSectionTransformer'
-    },
-    query: {
-      operation: 'getMarriageRegistrationSectionTransformer'
-    }
-  }
+  mapping: getSectionMapping('registration')
 } as ISerializedFormSection
 
 export const marriageDocumentType = {
